# EES-3976 - PNPM migration