--- a/.npmrc
+++ b/.npmrc
@@ -1,0 +1,16 @@
+engine-strict = true
+
+# feature parity with npm
+auto-install-peers=true
+
+# create a flat, non-symlinked node_modules structure 
+# (as the project doesn't play nice with symlinks). 
+# see https://pnpm.io/feature-comparison for more info
+shamefully-hoist=true
+
+# exclude react, react-dom & formik from hoisting due to 
+# the frontend project needing them present in the relative 
+# 'node_modules' directory in order to build properly.
+hoist-pattern[]=!react/*
+hoist-pattern[]=!react-dom/*
+hoist-pattern[]=!formik/*