--- a/docker/Dockerfile.public-frontend
+++ b/docker/Dockerfile.public-frontend
@@ -1,0 +1,40 @@
+FROM node:16.14.2-alpine AS builder
+
+WORKDIR /app
+
+COPY .npmrc .
+COPY package.json .
+COPY pnpm-lock.yaml .
+COPY pnpm-workspace.yaml .
+COPY src/explore-education-statistics-common/ ./src/explore-education-statistics-common/
+COPY src/explore-education-statistics-frontend/ ./src/explore-education-statistics-frontend/
+COPY src/tsconfig.json ./src/tsconfig.json
+
+RUN corepack enable
+
+RUN pnpm i
+RUN pnpm --filter=explore-education-statistics-frontend... install
+RUN pnpm --filter=explore-education-statistics-frontend build
+
+FROM node:16.14.2-alpine AS runner
+
+WORKDIR /usr/src/app
+
+COPY --from=builder /app/package.json .
+COPY --from=builder /app/pnpm-lock.yaml .
+COPY --from=builder /app/pnpm-workspace.yaml .
+COPY --from=builder /app/src/explore-education-statistics-common/ ./src/explore-education-statistics-common/
+COPY --from=builder /app/src/explore-education-statistics-frontend/ ./src/explore-education-statistics-frontend/
+
+RUN corepack enable
+
+RUN pnpm i
+RUN pnpm --filter=explore-education-statistics-frontend... --prod install --ignore-scripts
+
+WORKDIR /usr/src/app/src/explore-education-statistics-frontend
+
+ENV NODE_ENV=production
+EXPOSE 3000
+
+USER node
+CMD ["pnpm", "start"]