--- a/pnpm-workspace.yaml
+++ b/pnpm-workspace.yaml
@@ -1,0 +1,5 @@
+packages:
+  - 'src/explore-education-statistics-admin'
+  - 'src/explore-education-statistics-ckeditor'
+  - 'src/explore-education-statistics-common'
+  - 'src/explore-education-statistics-frontend'