--- a/src/explore-education-statistics-admin/babel.config.js
+++ b/src/explore-education-statistics-admin/babel.config.js
@@ -1,6 +1,12 @@
 module.exports = {
   presets: [['react-app', { flow: false, typescript: true }]],
-  plugins: ['@babel/plugin-proposal-class-properties'],
+  plugins: [
+    // needed to silence a warning about loose mode in unit tests
+    // and admin webpack logs
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
+    ['@babel/plugin-proposal-private-methods', { loose: true }],
+    ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
+  ],
   inputSourceMap: true,
   sourceMaps: true,
   env: {