--- a/src/explore-education-statistics-admin/src/pages/release/footnotes/components/IndicatorDetails.tsx
+++ b/src/explore-education-statistics-admin/src/pages/release/footnotes/components/IndicatorDetails.tsx
@@ -32,7 +32,7 @@ const IndicatorDetails = ({
               `${valuePath}.indicatorGroups.${indicatorGroupId}.selected`,
             );
             const hideGrouping = indicatorGroup.label === 'Default';
-            const indicators =
+            const indicators: string[] =
               get(
                 form.values,
                 `${valuePath}.indicatorGroups.${indicatorGroupId}.indicators`,
@@ -64,9 +64,7 @@ const IndicatorDetails = ({
                 >
                   {indicatorGroup.options.map(indicatorItem => {
                     const checked =
-                      (indicators &&
-                        indicators.includes(indicatorItem.value)) ||
-                      false;
+                      indicators.includes(indicatorItem.value) || false;
                     return (
                       <FormCheckbox
                         {...indicatorItem}