--- a/src/explore-education-statistics-common/src/components/form/FormCheckboxSearchSubGroups.tsx
+++ b/src/explore-education-statistics-common/src/components/form/FormCheckboxSearchSubGroups.tsx
@@ -125,9 +125,7 @@ const FormCheckboxSearchSubGroups = ({
 
   const handleAllGroupsChange: MouseEventHandler<HTMLButtonElement> = useCallback(
     event => {
-      if (onAllChange) {
-        onAllChange(event, isAllChecked, filteredOptions);
-      }
+      onAllChange?.(event, isAllChecked, filteredOptions);
     },
     [isAllChecked, onAllChange, filteredOptions],
   );
@@ -193,9 +191,7 @@ const FormCheckboxSearchSubGroups = ({
                   } subgroup options`
                 }
                 onAllChange={(event, checked) => {
-                  if (onSubGroupAllChange) {
-                    onSubGroupAllChange(event, checked, optionGroup.options);
-                  }
+                  onSubGroupAllChange?.(event, checked, optionGroup.options);
                 }}
                 hiddenText={`for ${optionGroup.legend}`}
               />
@@ -214,9 +210,7 @@ const FormCheckboxSearchSubGroups = ({
             value={value}
             selectAll
             onAllChange={(event, checked) => {
-              if (onSubGroupAllChange) {
-                onSubGroupAllChange(event, checked, filteredOptions[0].options);
-              }
+              onSubGroupAllChange?.(event, checked, filteredOptions[0].options);
             }}
           />
         )}