--- a/src/explore-education-statistics-common/src/components/form/FormTextSearchInput.tsx
+++ b/src/explore-education-statistics-common/src/components/form/FormTextSearchInput.tsx
@@ -23,9 +23,7 @@ const FormTextSearchInput = ({
 
   const [handleChange] = useDebouncedCallback(
     (event: ChangeEvent<HTMLInputElement>) => {
-      if (onChange) {
-        onChange(event);
-      }
+      onChange?.(event);
     },
     debounceTime,
   );