--- a/src/explore-education-statistics-frontend/src/middleware/withMethods.ts
+++ b/src/explore-education-statistics-frontend/src/middleware/withMethods.ts
@@ -19,9 +19,10 @@ export default function withMethods(config: MethodsConfig): NextApiHandler {
     const handler = config[method];
 
     if (!handler) {
-      return res.status(405).end();
+      res.status(405).end();
+      return;
     }
 
-    return handler(req, res);
+    handler(req, res);
   };
 }