… README.md
… azure-pipelines.dfe.yml
… infrastructure/templates/template.json
… lerna.json
… package.json
… pnpm-lock.yaml
… src/explore-education-statistics-admin/package.json
… src/explore-education-statistics-common/package.json
… src/explore-education-statistics-frontend/README.md
… src/explore-education-statistics-frontend/package.json