See: https://github.com/dfe-analytical-services/explore-education-statistics/pull/3782 for original PR review

This PR migrates the project from Lerna + NPM to PNPM & PNPM workspaces. 

Useful links:
* https://pnpm.io/installation
* https://github.com/pnpm/pnpm


## Getting started with PNPM

In order to set up the project & install dependencies for various packages, you would previously have to run two commands: `npm ci && npm run bootstrap`. With PNPM this command now simply turns into `pnpm i`.

`pnpm i` will install the root dependencies along with all the workspace packages defined in pnpm-workspace.yaml

When it comes to installing new dependencies, you simply modify the relevant package.json and run `pnpm i`

When it comes to running scripts you can choose to omit the run part of your command as PNPM can work just fine without it.
i.e.

### before 

```
npm run lint && npm run format 
```

### after 
```
pnpm lint && pnpm format 
```

Install PNPM & project dependencies

### install PNPM via corepack (https://github.com/nodejs/corepack)

```
corepack enable
```

If for some reason corepack doesn't work for you, try running the following in the root of the project
```
PNPM_VERSION=$(node -e "console.log(require('./package.json').engines.pnpm)")
curl -fsSL https://get.pnpm.io/install.sh | env PNPM_VERSION=$PNPM_VERSION sh -

```


### install dependencies 

```
pnpm i 
```
Useful links:

https://pnpm.io/installation
https://github.com/pnpm/pnpm

## Introduction

Due to the way PNPM symlinks dependencies and the stricter structure it provides out of the box (see https://pnpm.io/symlinked-node-modules-structure), we've had to enable the `shamefully-hoist` option in order to create a flat `node_module` structure.

In addition to this, we've excluded 3 packages from being hoisted to the root of the project. This is purely to ensure the frontend has `react`, `react-dom` and `formik` available in it's relative `node_modules` folder so that it can build properly. 

##  Differences between NPM and PNPM

Package managers such as Yarn or NPM use a flattened `node_modules` structure. The flattened dependency tree results in many issues such as:
 * The algorithm for flattening a dependency tree is complex
* Some packages have to be duplicated inside another project's `node_modules` folder 
* Modules have access to packages they do not depend on (also known as 'phantom' dependencies)

Let's go over these issues with an example. Let's say one project needs `express` as a dependency, and `express` has a dependency on a package named `debug`. This is how the structure of `node_modules` will look like with NPM: 

![Screenshot 2023-03-20 at 10 46 27](https://user-images.githubusercontent.com/55030296/226472664-d0648c33-53b3-453b-839e-ed8fd2ca8878.png)

At this stage, the code can import the debug package i.e. `require('debug')` without any issues whatsoever. We don't explicitly depend on this package via an entry in the `package.json` file. There are two big problems that would arise by doing this:

* Express updates its debug dependency with breaking changes
* Express does not depend on debug anymore 
* Your package now relies on 'phantom' packages making it harder to pinpoint what packages a project relies on

In both cases, our code will fail because it has an implicit dependency on debug. This is a big issue especially when we're working with large monorepos such as this one. It is much harder to track down particular dependencies that a given project uses. I've had to fix up a few issues around this area when migrating the project over to PNPM and it was pretty tedious + time-consuming.

### How does PNPM solve the above problems?
PNPM uses **hard links** and **symlinks** to maintain a **semistrict** `node_modules` structure. Let's take a look at the above example when using PNPM:

![Screenshot 2023-03-20 at 10 52 28](https://user-images.githubusercontent.com/55030296/226473822-e5982632-8cca-4225-b108-4b5b1faa633e.png)

The first difference you'll notice is that your code can no longer access the `debug` package because it is not directly under the root level `node_modules` directory. PNPM has created a special folder called `.pnpm` that containers hard links to all the modules.  If you look at `express@4.0.0`, the express module is a hard link to the global `pnpm-store` and a `debug` symbolic link to the `debug` hard link that also links to the global `pnpm-store`. 

Below is an image of the actual `pnpm-store`, which contains the packages. This is where all downloaded dependencies are maintained. When you download a dependency, PNPM first checks whether the dependency is present in this store or not. If it finds the dependency in the store, PNPM fetches it by creating a hard link.

<img width="360" alt="Screenshot 2023-03-20 at 21 57 56" src="https://user-images.githubusercontent.com/55030296/226474696-be31949b-3f13-4429-8350-588e28d63b9f.png">

Because of this approach, PNPM re-uses the same packages if they are already installed for another project. 

## PNPM benefits 

#### Disk space efficiency 
PNPM uses a content-addressable file system to store the packages and dependencies on disk. This means that the same package will not be duplicated. Even with different versions of the same package, PNPM is intelligent enough to reuse the maximum amount of code. i.e. if package version `1.0.0` has 500 files and version `1.0.1` has just one more file, then PNPM will not write 501 files for version `1.0.1`. Instead, it will create a hard link to the original 500 files and write just the new file. If you compare this with NPM, then version 1.0.0 will also be loaded, thus duplicating the original 500 files. For large monorepo projects, this can make a big difference (and make your `node_modules` folder GBs big 😅). 

#### Improved speed 

PNPM outperforms its peers several times over in most cases. Here's a screenshot taken from PNPM's website  

![alotta-files](https://user-images.githubusercontent.com/55030296/226476199-2a165313-43f8-4a55-a142-8dc5bc42a6f3.svg)

#### Security 

PNPM like Yarn, has a special file with the checksum of all the installed packages. This ensures that the integrity of all the installed packages is verified before their code is executed. 

In terms of unprivileged access, PNPM also outperforms both NPM & Yarn. In the case of NPM & Yarn, if package `A` depends on package `B`, and `B` depends on `C`, then `A` implicitly gets access to `C` even though `A` has not declared `C` as its dependency. This problem as you can imagine is intensified in a large monorepo setup. PNPM on the other hand, uses a different dependency resolution algorithm and different folder structure of `node_modules` that prevents illegal access to packages 


## PNPM CLI differences

npm command | Yarn command | pnpm equivalent
-- | -- | --
npm install | yarn | pnpm install
npm install [pkg] | yarn add [pkg] | pnpm add [pkg]
npm uninstall [pkg] | yarn remove [pkg] | pnpm remove [pkg]
npm update | yarn upgrade | pnpm update
npm list | yarn list | pnpm list
npm run [scriptName] | yarn [scriptName] | pnpm [scriptName]
npx [command] | yarn dlx [command] | pnpm dlx [command]
npm exec | yarn exec [commandName] | pnpm exec [commandName]
npm init [initializer] | yarn create [initializer] | pnpm create [initializer]


## CI speed improvements 
PNPM is supposedly 2 - 3x times faster than the alternatives (Yarn, NPM, Lerna, etc.). As can be seen in the screenshot below of the build pipeline (left = PNPM + PNPM workspaces, right = current implementation of NPM + Lerna) we're now saving lots of time spent in CI

![image (1)](https://user-images.githubusercontent.com/55030296/213530841-f5401381-767c-487c-a952-d74275cacdc1.png)

| Step         | NPM + Lerna     | PNPM + PNPM workspaces | difference |
|---------- |-------------- |------------|-----------------|
| Admin            |  11m 40s    |  8m 17s      | -3min 33s ✅
| Frontend       |   22m 42s  |   15m 46s   | -6min 54s ✅

Overall build time difference:
-6min 55seconds ✅

NPM build + PNPM build side by side for future reference (PNPM left, NPM right): 
![image (2)](https://user-images.githubusercontent.com/55030296/213533312-b42b1ed4-5386-4033-ad6d-7302a0b77118.png)

## Infrastructure changes

Due to standard Azure app services not supporting PNPM, we've had to build & deploy the public frontend via Azure's hybrid app-service model. This entails building and pushing a docker image to a container registry. The app service will then boot up the docker container to serve the project. Because of this, we've made some amends to how the public frontend is deployed:

* Created a 'Standard' tier Azure container registry to house docker images
* Added a step in the build pipeline to build + push new docker images on merges
* Adds three new environment variables to the frontend app service related to docker changes

## Relevant changes:

**Updated the build pipeline to Install and use PNPM**
* Since Azure seems to not have a `PNPM` task, I've resorted to just using bash to run relevant PNPM commands.  
* Moved commonly used version specs (for things like Dotnet, Node, etc.) to variables for better organization
* Added appropriate spacing so it's easier to read

**Updated relevant READMEs** 
* Updated public frontend README and removed old information
* Updated main README due to changes in commands 

**General cleanup** 
* Removed check-node-version package as PNPM natively protects against wrong node/PNPM versions being used with the engines field in package.json and engine-strict in .npmrc
* Fixed type errors in `FilterGroupDetails.tsx` and `IndicatorDetails.tsx` that seem to have cropped up
* Fixes flaky (at least for myself locally) `Modal.test.tsx` unit test
* Fixes eslint warning in `MapBlockInternal.tsx`
* Upgrades `postcss-preset-env` to v7+ to avoid a spammy warning in the admin webpack logs: 

**Upgraded admin babel packages** 
* Upgraded admin babel packages to stop a problem occurring in CK editor whereby the editor simply wouldn't work and would crash whenever you navigated to parts of the admin app that rendered an editor. As part of this, amends needed to be made to the Admin's babel config. This was due to log spam in unit tests and when the admin initially started up. See https://stackoverflow.com/questions/67770240/webpacker-babel-spam-removal-message-not-clear-enough for an example of the warnings produced